Improve the checking procedure and emit better messages; v0.1.4
When the output of a service would stop conforming to the spec,
oftentimes the checker script would fail with a CRITICAL message that
doesn't really say much about the failure. This commit improves the
checking procedure, and consequently the error messages. Concretely:

- Add a default title to each x-ample. The code base assumes each
  x-maple has a title, so better to provide one.
- Add a default request and response if any of them is missing.
- Handle None values appropriately. If the spec has a None value, the
  checker script cannot discern what the correct value should be, so
  bail out in that case. Similarly, if the returned value is None,
  assume it means "empty value".
- For dicts and lists, ensure the values are of the same type. If the
  spec expects a dict or a list, do not assume that the value is of that
  same type, but check it before descending into their elements.
- For dicts, do not assume all expected fields exist. Instead, check
  that they do before following the path.
- For lists, check the expected amount of elements exist. If the spec
  contains only one element, use it for comparison against all of the
  returned values. Otherwise, check that the number of elements is
  equal.

Each of the listed improvements also sets the WARNING level and an
appropriate message if any of the checks fails. Additionally, the
checking procedure will not be stopped if any of these checks fail,
warnings will be accumulated so that the user can see all of them at
once.

Bug: T150560
Change-Id: Ide292c7ae8ced694e97fd7cde84ab65c28a69185