Convert to non native debian package
Having service-checker as a native debian package is problematic in
multiple ways. Both conceptually as only related to debian internals
packages should be marked as native e.g. apt, dpkg as well as
functionally as in our case distributing the package for multiple
versions requires a version bump instead of a debian revision bump.
Which in turns causes major confusion as the impression that e.g. jessie
vs stretch run different versions of the software with different
features

Change-Id: Iea5afd9aafd548aaa8391ace6917cdf861779382